(Feat(OAuth2AuthenticationSuccessHandler) Enables OAuth2 functionality by implementing a custom onAuthenticationSuccess method.

+++ src/main/java/com/planify/planify/service/FriendshipServiceImpl.java
@@ -54,6 +54,9 @@ public class FriendshipServiceImpl implements FriendshipService {
 
     @Override
     public void deleteFriendship(long friendshipId) {
+        if (!friendshipRepository.existsById(friendshipId)) {
+            throw new IllegalStateException("Friendship does not exist");
+        }
         friendshipRepository.deleteById(friendshipId);
 
     }
